Add basic API endpoints (stubbed)
The initial version of cliff just needs register and send endpoints, and
won't support clearing notifications on other devices when viewed. In
the future, more endpoints will be necessary to support this behavior.

Also note that the APNs part of these endpoints are stubbed. The
handlers correctly fetch the Tailscale information for the node making
the request (and log it), but devices are not registered and
notifications are not sent.